salva tudo

+++ A3 Teste/produto.cpp
@@ -1,6 +1,6 @@
 #include "produto.h"
 
-Produto::Produto(int _id,const string& _nome, float _preco, const string& _tamanho, const string& _cor, const string& _marca, int _quantidade, const string& _tipo) : id(_id), nome(_nome), preco(_preco), tamanho(_tamanho), cor(_cor), marca(_marca), quantidade(_quantidade), tipo(_tipo) {
+Produto::Produto(int _id,const string& _nome, double _preco, const string& _tamanho, const string& _cor, const string& _marca, int _quantidade, const string& _tipo) : id(_id), nome(_nome), preco(_preco), tamanho(_tamanho), cor(_cor), marca(_marca), quantidade(_quantidade), tipo(_tipo) {
 }
 
 int Produto::getId() const
@@ -12,7 +12,7 @@ string Produto::getNome() const {
 	return nome;
 }
 
-float Produto::getPreco() const {
+double Produto::getPreco() const {
 	return preco;
 }
