+++ A3 Teste/produto.h
@@ -11,7 +11,7 @@ class Produto {
 
 	int id;
 	string nome;
-	float preco;
+	double preco;
 	string tamanho;
 	string cor;
 	string marca;
@@ -20,11 +20,11 @@ class Produto {
 
 	public:
 
-	Produto(int id,const string& _nome, float _preco, const string& _tamanho, const string& _cor, const string& _marca, int quantidade, const string& tipo);
+	Produto(int id,const string& _nome, double _preco, const string& _tamanho, const string& _cor, const string& _marca, int quantidade, const string& tipo);
 
 	int getId() const;
 	string getNome() const;
-	float getPreco() const;
+	double getPreco() const;
 	string getTamanho() const;
 	string getCor() const;
 	string getMarca() const;
